Make delete great again
