+++ src/components/absences/delete/DeleteAbsenceModal.tsx
@@ -1,0 +1,98 @@
+import {
+  Button,
+  Modal,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  ModalOverlay,
+  Text,
+  useToast,
+} from '@chakra-ui/react';
+import { useUserData } from '@hooks/useUserData';
+import { Role } from '@utils/types';
+import { useState } from 'react';
+
+interface DeleteAbsenceModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  absenceId: number;
+  onDelete: (absenceId: number) => Promise<void>;
+}
+
+const DeleteAbsenceModal: React.FC<DeleteAbsenceModalProps> = ({
+  isOpen,
+  onClose,
+  absenceId,
+  onDelete,
+}) => {
+  const [isDeleting, setIsDeleting] = useState(false);
+  const toast = useToast();
+  const userData = useUserData();
+  const isUserAdmin = userData.role === Role.ADMIN;
+
+  const handleDeleteConfirm = async () => {
+    try {
+      setIsDeleting(true);
+      const response = await fetch(`/api/deleteAbsence`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          isUserAdmin,
+          absenceId,
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to delete absence');
+      }
+
+      toast({
+        title: 'Absence deleted',
+        description: 'The absence has been successfully deleted.',
+        status: 'success',
+        duration: 5000,
+        isClosable: true,
+      });
+
+      await onDelete(absenceId);
+      onClose();
+    } catch (error) {
+      toast({
+        title: 'Error',
+        description: error.message || 'Failed to delete absence',
+        status: 'error',
+        duration: 5000,
+        isClosable: true,
+      });
+    } finally {
+      setIsDeleting(false);
+    }
+  };
+
+  return (
+    <Modal isOpen={isOpen} onClose={onClose} isCentered>
+      <ModalOverlay />
+      <ModalContent>
+        <ModalHeader>Delete Absence</ModalHeader>
+        <ModalCloseButton />
+        <ModalBody>
+          <Text>Are you sure you want to delete this absence?</Text>
+        </ModalBody>
+        <ModalFooter>
+          <Button onClick={onClose} mr={3}>
+            Cancel
+          </Button>
+          <Button onClick={handleDeleteConfirm} isLoading={isDeleting}>
+            Delete
+          </Button>
+        </ModalFooter>
+      </ModalContent>
+    </Modal>
+  );
+};
+
+export default DeleteAbsenceModal;
